+++ app/(tabs)/_layout.tsx
@@ -19,7 +19,7 @@ export default function TabLayout() {
       }}
     >
       <Tabs.Screen
-        name="index"
+        name="historico"
         options={{
           title: "History",
           tabBarIcon: ({ color, focused }) => (
@@ -31,7 +31,7 @@ export default function TabLayout() {
         }}
       />
       <Tabs.Screen
-        name="explore"
+        name="index"
         options={{
           title: "Home",
           tabBarIcon: ({ color, focused }) => (
