+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "blank-typescript",
+  "name": "controle-ponto",
   "main": "expo-router/entry",
   "version": "1.0.0",
   "scripts": {
